Fix error message for external node syntax used, no leading spaces and simplify function call.
In Texinfo::Report, allow not to output error messages, if
$self->{'ignore_notice'} is set.

Add an explanation string to the _convert call, mostly to have
more readable cluck traces...

No error messages in contexts of multiple formatting of a texinfo
command.

Handle SIMPLE_MENU and preformat menu_entry and menu_comment if
in_preformatted.


git-svn-id: svn://127.0.0.1/svn_repo/texinfo/trunk@3909 39fee189-59d7-47db-b5d4-205258b72aed