Implement index subentries in texinfo.tex.
* doc/texinfo.tex (\splitindexentry): New macro.
(\dosubindwrite): Call it to get the index text split up into
primary entry and possible secondary and further entries
separated by @subentry, as well as the index sort key.  For one
subentry only, this is the same format as was output in the past
and recognized by the C implementation of texindex.

This feature is experimental.