Remove apparently unused code.
* doc/texinfo.tex (\begindoublecolumns): Remove code that is
supposed to account for the output routine in
\begindoublecolumns being called twice in a row.  It checks
whether \partialpage is void, but it appears that \partialpage
will always be void.  \doublecolumnout, \balancecolumns and
\pagesofar all clear \partialpage with \unvbox, which I believe
covers all the cases.  I was unable to get this code to run with
various tests.  Remove references to long-lost test files.
(Removed code introduced on Jan 14 16:53:43 1999 according to
ChangeLog entry.)