Support links between locally-installed manuals.
* js/info.js (external_manual_url_p): New function.
(fix_links): If a link to an external manual, open link in the
same browser tab.
(on_click): No special handling of links to external manuals.

This is intended to support links between locally-installed HTML
manuals.  For this to work, links to other manuals must begin
with "../", so htmlxref.cnf would need to be overridden.