Optionally use backslash as index escape character.
* doc/texinfo.tex
(\doindwrite): If the user gave '@set txiindexescapeisbackslash',
then use \ as the escape character in the index file.
(\splitindexentry): Use \string to output \indexlbrace etc.
(\printindex): Under the same condition, read the sorted index
file with backslash as an escape character.  Expand the advice
on how to print the indices properly.