Fix parsetexi macro handling.
* tp/Texinfo/XS/parsetexi/macro.c (handle_macro): For a macro
taking a single argument, if the rest of the line is empty,
leave the newline on the line.  Otherwise, a command on the next
line would not be recognized as occurring at the beginning of
the line.  This led to errors with as.texi from binutils,
reported by Hans-Bernhard Bröker.

* tp/t/60macro.t (include_after_empty_line_arg): New test.