Remove unused macro.
* doc\texinfo.tex (\pagealignmacro): Remove macro.  It was not
used except in \bye.
(\bye): Use \chappager instead of \pagealignmacro.  Presumably
we do not need an extra blank page at the very end of the
document if "@setchapternewpage odd" to make sure that we end on
an even page.
(\startcontents): Remove a comment.  The comment credited
Torbjorn Granlund <tege@matematik.su.se> for the
\contentsalignmacro, presumably added 1996-08-03.