Optional parameter to @example
* tp/Texinfo/Common.pm (%block_commands),
* tp/Texinfo/XS/parsetexi/command_data.txt: Specify @example to
take 1 argument.
* tp/Texinfo/Convert/HTML.pm (_convert_preformatted_command):
If 'example' takes an argument, pass it as an extra argument
to _attribute_class.
(_attribute_class): Take an extra argument for extra classes
to be added to the element.

This could be used as e.g. "@example cplusplus" to allow
post-processing the HTML output to syntax-highlight C++ code.